Use middleware to extract token from route
