+++ WebApplication/Middleware/RouteTokenHandler.cs
@@ -1,4 +1,4 @@
-/////////////////////////////////////////////////////////////////////
+﻿/////////////////////////////////////////////////////////////////////
 // Copyright (c) Autodesk, Inc. All rights reserved
 // Written by Forge Design Automation team for Inventor
 //
@@ -17,43 +17,48 @@
 /////////////////////////////////////////////////////////////////////
 
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Http;
-using Microsoft.Net.Http.Headers;
+using Microsoft.AspNetCore.Routing;
+using Microsoft.Extensions.Logging;
 using WebApplication.State;
 
 namespace WebApplication.Middleware
 {
     /// <summary>
-    /// Middleware to extract access token from HTTP headers.
+    /// Middleware to extract access token from route parameters.
     /// </summary>
-    public class TokenHandler
+    public class RouteTokenHandler
     {
-        private const string BearerPrefix = "Bearer ";
-
         private readonly RequestDelegate _next;
 
-        public TokenHandler(RequestDelegate next)
+        public RouteTokenHandler(RequestDelegate next)
         {
             _next = next;
         }
 
-        public async Task InvokeAsync(HttpContext context, UserResolver resolver)
+        public async Task InvokeAsync(HttpContext context, UserResolver resolver, ILogger<RouteTokenHandler> logger)
         {
-            while (context.Request.Headers.TryGetValue(HeaderNames.Authorization, out var values))
+            string token = context.GetRouteValue("token") as string; // IMPORTANT: parameter name must be in sync with route definition
+            if (!string.IsNullOrEmpty(token))
             {
-                var headerValue = values[0];
-                if (headerValue.Length <= BearerPrefix.Length) break;
-                if (! headerValue.StartsWith(BearerPrefix)) break;
-
-                string token = headerValue.Substring(BearerPrefix.Length);
-                if (string.IsNullOrEmpty(token)) break;
-
+                logger.LogInformation("Extracted token from route");
                 resolver.Token = token;
-                break;
             }
 
             // Call the next delegate/middleware in the pipeline
             await _next(context);
+        }
+    }
+
+    /// <summary>
+    /// Special pipeline to extract tokens from route values
+    /// </summary>
+    public class RouteTokenPipeline
+    {
+        public void Configure(IApplicationBuilder applicationBuilder)
+        {
+            applicationBuilder.UseMiddleware<RouteTokenHandler>();
         }
     }
 }
